+++ R/api_v2_responses.R
@@ -1,11 +1,12 @@
 # Handling responses ####
-parsing <- function(x, expansions, fields, call = caller_env()) {
+parsing <- function(x, expansions, fields) {
   if (!is_logical(x)) {
-    abort("parse should be either TRUE or FALSE", call = call)
+    abort("parse should be either TRUE or FALSE", call = current_call())
   }
   if (isTRUE(x) && (!is.null(expansions) || !is.null(fields))) {
     abort(c("Not yet implemented!",
-            i = "Stay tuned for further updates or use `parse = FALSE`"))
+            i = "Stay tuned for further updates or use `parse = FALSE`"),
+          call = current_call())
   }
 }
 
@@ -17,6 +18,8 @@ list_minus <- function(l, minus) {
 # Pagination should be consistent across API v2
 # <https://developer.twitter.com/en/docs/twitter-api/pagination>
 pagination <- function(req, n_pages, count, verbose = TRUE) {
+  # To store the token at the right place: see ?httr2::oauth_cache_path
+  withr::local_envvar(HTTR2_OAUTH_CACHE = auth_path())
   if (is.infinite(n_pages)) {
     n_pages <- 8
   }
@@ -24,7 +27,7 @@ pagination <- function(req, n_pages, count, verbose = TRUE) {
   tmp <- tempfile("rtweet_tmp", fileext = ".rds")
 
   all_results <- vector("list", length = n_pages)
-  resp <- httr2::req_perform(req)
+  resp <- httr2::req_perform(req, error_call = sys.call(1))
   x0 <- resp(resp)
   all_results[[1]] <- x0
 
@@ -83,7 +86,7 @@ resp <- function(x, ...) {
   class(out) <- c("Twitter_resp", class(out))
 
   if (has_name_(out, "errors")) {
-    abort(req_errors(out), call = NULL)
+    abort(req_errors(out), call = current_call())
   }
 
   if (has_name_(out, "meta")) {
@@ -103,7 +106,7 @@ resp <- function(x, ...) {
     }
 
     if (nrow(rest) > 1) {
-      abort("Please check", call = call)
+      abort("Please check", call = current_call())
     }
     out$meta <- rest
   }