+++ R/client.R
@@ -71,7 +71,8 @@ default_client <- function(client_id = NULL, client_secret = NULL) {
   if (is.null(client_id) && is.null(client_secret)) {
     # The sysdat file is in #./R and loaded automagically
     abort(c("The default rtweet client is no longer authorized.",
-            i = "You'll need to register as developer in order to use the Twitter API."))
+            i = "You'll need to register as developer in order to use the Twitter API."),
+          call = caller_env())
     client_id <- decrypt(sysdat$e914c55d2f)
     client_secret <- decrypt(sysdat$d5571d4003)
   } else {
@@ -131,18 +132,16 @@ client_list <- function() {
 
 find_client <- function(client = NULL) {
   if (is.null(client)) {
-    if (is_developing()) {
-      load_client("academic_dev") %||% no_client()
-    } else{
-      default_cached_client()
-    }
+    no_client()
   } else if (is_client(client)) {
-    client
+    client <- client
   } else if (is_string(client)) {
-    load_client(client)
+    client <- load_client(client)
   } else {
-    abort("Unrecognised input to `client`")
+    abort("Unrecognised input to `client`", call = current_call())
   }
+  inform(paste("Using client", client$name))
+  client
 }
 
 
@@ -154,7 +153,8 @@ load_client <- function(client_name) {
     path <- client_path(paste0(client_name, ".rds"))
   }
   if (!file.exists(path) && !is_developing()) {
-    abort(paste0("Can't find saved client with name '", client_name, "'"))
+    abort(paste0("Can't find saved client with name '", client_name, "'"),
+          call = current_call())
   } else if (!file.exists(path)) {
     return(NULL)
   }
@@ -169,7 +169,7 @@ no_client <- function(call = caller_env()) {
   if (is_testing()) {
     testthat::skip("Client not available")
   } else {
-    abort("Could not find client", call = call)
+    abort(c("Could not find client", i = "Find which you have with `client_list()`"), call = sys.call(1))
   }
 }
 
@@ -185,17 +185,20 @@ no_client <- function(call = caller_env()) {
 #' @seealso scopes
 #' @export
 #' @examples
+#' \donttest{
 #' if (interactive()) {
 #'   rtweet_client()
 #' }
+#' }
 rtweet_client <- function(client_id, client_secret,
                           app, scopes = NULL) {
-  if (missing(client_id) && missing(client_secret) ) {
-    dc <- default_client()
-    client_id <- dc[1]
-    client_secret <- dc[2]
-    app <- "rtweet"
-  }
+  # The default rtweet client no longer works
+  # if (missing(client_id) && missing(client_secret) ) {
+  #   dc <- default_client()
+  #   client_id <- dc[1]
+  #   client_secret <- dc[2]
+  #   app <- "rtweet"
+  # }
 
   if (missing(client_id) && interactive()) {
     client_id <- ask_pass("client ID key")
@@ -218,7 +221,6 @@ rtweet_client <- function(client_id, client_secret,
     token_url = "https://api.twitter.com/2/oauth2/token",
     auth = "header",
     name = app)
-  attr(client, "app") <- app
   attr(client, "scopes") <- scopes
   client
 }
@@ -248,7 +250,8 @@ client_setup_default <- function() {
     inform("Using default client available.")
   } else {
     abort(c("The default rtweet client is no longer authorized.",
-            i = "You'll need to register as developer in order to use the Twitter API."))
+            i = "You'll need to register as developer in order to use the Twitter API."),
+          call = current_call())
     client <- rtweet_client(decrypt(sysdat$DYKcJfBkgMnGveI),
                             decrypt(sysdat$MRsnZtaKXqGYHju),
                             app = "rtweet")
@@ -259,7 +262,7 @@ client_setup_default <- function() {
 }
 
 
-client_scopes <- function(client, call = caller_env()) {
+client_scopes <- function(client) {
   if (!is_client(client)) {
     abort(c("Missing client.",
             ">" = "Check the vignette('auth', 'rtweet')",
@@ -271,4 +274,16 @@ client_scopes <- function(client, call = caller_env()) {
 client_app <- function(client) {
   stopifnot(is_client(client))
   attr(client, "app", exact = TRUE)
+}
+
+#' Clean clients
+#'
+#' Removes the rtweet clients directory.
+#' @returns Either TRUE if folder could be deleted or FALSE.
+#' @export
+#' @examples
+#' client_clean()
+client_clean <- function() {
+  out <- unlink(client_path(), recursive = TRUE, force = TRUE)
+  invisible(out == 0L)
 }