+++ R/post-user.R
@@ -1,5 +1,6 @@
 #' Follows target Twitter user.
 #'
+#' `r lifecycle::badge("deprecated")`
 #' @inheritParams get_timeline
 #' @param destroy Logical indicating whether to post (add) or
 #'   remove (delete) target tweet as favorite.
@@ -11,12 +12,8 @@
 #' @param retweets Logical indicating whether to enable retweets
 #'   for target user. Defaults to true.
 #' @aliases follow_user
-#' @examples
-#' if (auth_has_default()) {
-#'     post_follow("_R_Foundation")
-#'     post_follow("rtweet", mute = TRUE) # Mute
-#' }
 #' @family post
+#' @seealso [`rtweet-deprecated`]
 #' @export
 #' @references
 #' Update: <https://developer.twitter.com/en/docs/twitter-api/v1/accounts-and-users/follow-search-get-users/api-reference/post-friendships-update>